# backend/api/gen/go/oa/v1/workflow_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             (unknown)
// source: oa/v1/workflow.proto

package oav1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
	emptypb "google.golang.org/protobuf/types/known/emptypb"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationWorkflowServiceAuditTask = "/oa.v1.WorkflowService/AuditTask"
const OperationWorkflowServiceCreateWorkflowDefinition = "/oa.v1.WorkflowService/CreateWorkflowDefinition"
const OperationWorkflowServiceGetMyTasks = "/oa.v1.WorkflowService/GetMyTasks"
const OperationWorkflowServiceGetTask = "/oa.v1.WorkflowService/GetTask"
const OperationWorkflowServiceListWorkflowDefinition = "/oa.v1.WorkflowService/ListWorkflowDefinition"
const OperationWorkflowServiceSubmitApply = "/oa.v1.WorkflowService/SubmitApply"

type WorkflowServiceHTTPServer interface {
	// AuditTask 审批/驳回/转办当前待办任务。状态机在此推进实例状态、生成下一节点任务或终结实例。
	AuditTask(context.Context, *AuditTaskRequest) (*emptypb.Empty, error)
	// CreateWorkflowDefinition 创建工作流定义（管理端）。
	CreateWorkflowDefinition(context.Context, *CreateWorkflowDefinitionRequest) (*WorkflowDefinition, error)
	// GetMyTasks 获取当前用户的待办 / 已办 / 我的申请列表。
	GetMyTasks(context.Context, *GetMyTasksRequest) (*GetMyTasksResponse, error)
	// GetTask 获取单个待办任务详情：申请标题、申请表单数据、该实例的审批日志轨迹。
	//
	// 鉴权：仅当前指派审批人可查其处于 PENDING 的任务（与 AuditTask 同款授权，
	// IDEQ + AssigneeUserIDEQ(caller) + TaskStatusEQ(PENDING)，见
	// WorkflowTaskRepo.GetDetailByAssignee）。其余字段（definition_id /
	// current_node_index / tenant_id 等）不投影，对齐 MyTaskItem 的最小披露原则。
	GetTask(context.Context, *GetTaskRequest) (*GetTaskResponse, error)
	// ListWorkflowDefinition 查询工作流定义列表（管理端）。
	ListWorkflowDefinition(context.Context, *ListWorkflowDefinitionRequest) (*ListWorkflowDefinitionResponse, error)
	// SubmitApply 提交申请：发起一个工作流实例，按定义的首节点生成首条待办任务并通知该审批人。
	SubmitApply(context.Context, *SubmitApplyRequest) (*SubmitApplyResponse, error)
}

func RegisterWorkflowServiceHTTPServer(s *http.Server, srv WorkflowServiceHTTPServer) {
	r := s.Route("/")
	r.POST("/admin/v1/oa/workflow/definitions", _WorkflowService_CreateWorkflowDefinition0_HTTP_Handler(srv))
	r.GET("/admin/v1/oa/workflow/definitions", _WorkflowService_ListWorkflowDefinition0_HTTP_Handler(srv))
	r.POST("/admin/v1/oa/workflow/apply", _WorkflowService_SubmitApply0_HTTP_Handler(srv))
	r.POST("/admin/v1/oa/workflow/tasks/{task_id}/audit", _WorkflowService_AuditTask0_HTTP_Handler(srv))
	r.GET("/admin/v1/oa/workflow/my-tasks", _WorkflowService_GetMyTasks0_HTTP_Handler(srv))
	r.GET("/admin/v1/oa/workflow/tasks/{task_id}", _WorkflowService_GetTask0_HTTP_Handler(srv))
}

func _WorkflowService_CreateWorkflowDefinition0_HTTP_Handler(srv WorkflowServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateWorkflowDefinitionRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationWorkflowServiceCreateWorkflowDefinition)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateWorkflowDefinition(ctx, req.(*CreateWorkflowDefinitionRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*WorkflowDefinition)
		return ctx.Result(200, reply)
	}
}

func _WorkflowService_ListWorkflowDefinition0_HTTP_Handler(srv WorkflowServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListWorkflowDefinitionRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationWorkflowServiceListWorkflowDefinition)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListWorkflowDefinition(ctx, req.(*ListWorkflowDefinitionRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListWorkflowDefinitionResponse)
		return ctx.Result(200, reply)
	}
}

func _WorkflowService_SubmitApply0_HTTP_Handler(srv WorkflowServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in SubmitApplyRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationWorkflowServiceSubmitApply)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SubmitApply(ctx, req.(*SubmitApplyRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*SubmitApplyResponse)
		return ctx.Result(200, reply)
	}
}

func _WorkflowService_AuditTask0_HTTP_Handler(srv WorkflowServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in AuditTaskRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationWorkflowServiceAuditTask)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.AuditTask(ctx, req.(*AuditTaskRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _WorkflowService_GetMyTasks0_HTTP_Handler(srv WorkflowServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetMyTasksRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationWorkflowServiceGetMyTasks)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetMyTasks(ctx, req.(*GetMyTasksRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetMyTasksResponse)
		return ctx.Result(200, reply)
	}
}

func _WorkflowService_GetTask0_HTTP_Handler(srv WorkflowServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetTaskRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationWorkflowServiceGetTask)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetTask(ctx, req.(*GetTaskRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetTaskResponse)
		return ctx.Result(200, reply)
	}
}

type WorkflowServiceHTTPClient interface {
	// AuditTask 审批/驳回/转办当前待办任务。状态机在此推进实例状态、生成下一节点任务或终结实例。
	AuditTask(ctx context.Context, req *AuditTaskRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	// CreateWorkflowDefinition 创建工作流定义（管理端）。
	CreateWorkflowDefinition(ctx context.Context, req *CreateWorkflowDefinitionRequest, opts ...http.CallOption) (rsp *WorkflowDefinition, err error)
	// GetMyTasks 获取当前用户的待办 / 已办 / 我的申请列表。
	GetMyTasks(ctx context.Context, req *GetMyTasksRequest, opts ...http.CallOption) (rsp *GetMyTasksResponse, err error)
	// GetTask 获取单个待办任务详情：申请标题、申请表单数据、该实例的审批日志轨迹。
	//
	// 鉴权：仅当前指派审批人可查其处于 PENDING 的任务（与 AuditTask 同款授权，
	// IDEQ + AssigneeUserIDEQ(caller) + TaskStatusEQ(PENDING)，见
	// WorkflowTaskRepo.GetDetailByAssignee）。其余字段（definition_id /
	// current_node_index / tenant_id 等）不投影，对齐 MyTaskItem 的最小披露原则。
	GetTask(ctx context.Context, req *GetTaskRequest, opts ...http.CallOption) (rsp *GetTaskResponse, err error)
	// ListWorkflowDefinition 查询工作流定义列表（管理端）。
	ListWorkflowDefinition(ctx context.Context, req *ListWorkflowDefinitionRequest, opts ...http.CallOption) (rsp *ListWorkflowDefinitionResponse, err error)
	// SubmitApply 提交申请：发起一个工作流实例，按定义的首节点生成首条待办任务并通知该审批人。
	SubmitApply(ctx context.Context, req *SubmitApplyRequest, opts ...http.CallOption) (rsp *SubmitApplyResponse, err error)
}

type WorkflowServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewWorkflowServiceHTTPClient(client *http.Client) WorkflowServiceHTTPClient {
	return &WorkflowServiceHTTPClientImpl{client}
}

// AuditTask 审批/驳回/转办当前待办任务。状态机在此推进实例状态、生成下一节点任务或终结实例。
func (c *WorkflowServiceHTTPClientImpl) AuditTask(ctx context.Context, in *AuditTaskRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/admin/v1/oa/workflow/tasks/{task_id}/audit"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationWorkflowServiceAuditTask))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// CreateWorkflowDefinition 创建工作流定义（管理端）。
func (c *WorkflowServiceHTTPClientImpl) CreateWorkflowDefinition(ctx context.Context, in *CreateWorkflowDefinitionRequest, opts ...http.CallOption) (*WorkflowDefinition, error) {
	var out WorkflowDefinition
	pattern := "/admin/v1/oa/workflow/definitions"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationWorkflowServiceCreateWorkflowDefinition))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetMyTasks 获取当前用户的待办 / 已办 / 我的申请列表。
func (c *WorkflowServiceHTTPClientImpl) GetMyTasks(ctx context.Context, in *GetMyTasksRequest, opts ...http.CallOption) (*GetMyTasksResponse, error) {
	var out GetMyTasksResponse
	pattern := "/admin/v1/oa/workflow/my-tasks"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationWorkflowServiceGetMyTasks))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetTask 获取单个待办任务详情：申请标题、申请表单数据、该实例的审批日志轨迹。
//
// 鉴权：仅当前指派审批人可查其处于 PENDING 的任务（与 AuditTask 同款授权，
// IDEQ + AssigneeUserIDEQ(caller) + TaskStatusEQ(PENDING)，见
// WorkflowTaskRepo.GetDetailByAssignee）。其余字段（definition_id /
// current_node_index / tenant_id 等）不投影，对齐 MyTaskItem 的最小披露原则。
func (c *WorkflowServiceHTTPClientImpl) GetTask(ctx context.Context, in *GetTaskRequest, opts ...http.CallOption) (*GetTaskResponse, error) {
	var out GetTaskResponse
	pattern := "/admin/v1/oa/workflow/tasks/{task_id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationWorkflowServiceGetTask))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ListWorkflowDefinition 查询工作流定义列表（管理端）。
func (c *WorkflowServiceHTTPClientImpl) ListWorkflowDefinition(ctx context.Context, in *ListWorkflowDefinitionRequest, opts ...http.CallOption) (*ListWorkflowDefinitionResponse, error) {
	var out ListWorkflowDefinitionResponse
	pattern := "/admin/v1/oa/workflow/definitions"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationWorkflowServiceListWorkflowDefinition))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// SubmitApply 提交申请：发起一个工作流实例，按定义的首节点生成首条待办任务并通知该审批人。
func (c *WorkflowServiceHTTPClientImpl) SubmitApply(ctx context.Context, in *SubmitApplyRequest, opts ...http.CallOption) (*SubmitApplyResponse, error) {
	var out SubmitApplyResponse
	pattern := "/admin/v1/oa/workflow/apply"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationWorkflowServiceSubmitApply))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}