# backend/app/oa/service/cmd/server/assets/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /admin/v1/oa/workflow/apply:
        post:
            tags:
                - WorkflowService
            description: 提交申请：发起一个工作流实例，按定义的首节点生成首条待办任务并通知该审批人。
            operationId: WorkflowService_SubmitApply
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/SubmitApplyRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SubmitApplyResponse'
    /admin/v1/oa/workflow/definitions:
        get:
            tags:
                - WorkflowService
            description: 查询工作流定义列表（管理端）。
            operationId: WorkflowService_ListWorkflowDefinition
            parameters:
                - name: paging.page
                  in: query
                  description: 当前页码（从1开始，默认1）
                  schema:
                    type: integer
                    format: uint32
                - name: paging.pageSize
                  in: query
                  description: 每页条数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: paging.offset
                  in: query
                  description: 跳过的记录数（从0开始，默认0）
                  schema:
                    type: string
                - name: paging.limit
                  in: query
                  description: 最多返回的记录数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: paging.token
                  in: query
                  description: 上一页最后一条记录的游标（如ID/时间戳+ID，首次请求为空）
                  schema:
                    type: string
                - name: paging.noPaging
                  in: query
                  description: 是否不分页，如果为true，则page和pageSize参数无效。
                  schema:
                    type: boolean
                - name: paging.query
                  in: query
                  description: JSON字符串过滤条件，基础语法：{"field1":"val1", "field2___icontains":"val2"}，具体请参见：https://github.com/tx7do/go-crud/tree/main/pagination/filter/README.md
                  schema:
                    type: string
                - name: paging.filter
                  in: query
                  description: Google AIP规范字符串过滤条件
                  schema:
                    type: string
                - name: paging.filterExpr.type
                  in: query
                  description: 过滤表达式类型
                  schema:
                    enum:
                        - EXPR_TYPE_UNSPECIFIED
                        - AND
                        - OR
                    type: string
                    format: enum
                - name: paging.orderBy
                  in: query
                  description: 排序条件
                  schema:
                    type: string
                - name: paging.fieldMask
                  in: query
                  description: 字段掩码，其作用为SELECT中的字段，其语法为使用逗号分隔字段名，例如：id,realName,userName。如果为空则选中所有字段，即SELECT *。
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListWorkflowDefinitionResponse'
        post:
            tags:
                - WorkflowService
            description: 创建工作流定义（管理端）。
            operationId: WorkflowService_CreateWorkflowDefinition
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateWorkflowDefinitionRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/WorkflowDefinition'
    /admin/v1/oa/workflow/my-tasks:
        get:
            tags:
                - WorkflowService
            description: 获取当前用户的待办 / 已办 / 我的申请列表。
            operationId: WorkflowService_GetMyTasks
            parameters:
                - name: listType
                  in: query
                  schema:
                    enum:
                        - PENDING
                        - DONE
                        - SUBMITTED
                    type: string
                    format: enum
                - name: paging.page
                  in: query
                  description: 当前页码（从1开始，默认1）
                  schema:
                    type: integer
                    format: uint32
                - name: paging.pageSize
                  in: query
                  description: 每页条数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: paging.offset
                  in: query
                  description: 跳过的记录数（从0开始，默认0）
                  schema:
                    type: string
                - name: paging.limit
                  in: query
                  description: 最多返回的记录数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: paging.token
                  in: query
                  description: 上一页最后一条记录的游标（如ID/时间戳+ID，首次请求为空）
                  schema:
                    type: string
                - name: paging.noPaging
                  in: query
                  description: 是否不分页，如果为true，则page和pageSize参数无效。
                  schema:
                    type: boolean
                - name: paging.query
                  in: query
                  description: JSON字符串过滤条件，基础语法：{"field1":"val1", "field2___icontains":"val2"}，具体请参见：https://github.com/tx7do/go-crud/tree/main/pagination/filter/README.md
                  schema:
                    type: string
                - name: paging.filter
                  in: query
                  description: Google AIP规范字符串过滤条件
                  schema:
                    type: string
                - name: paging.filterExpr.type
                  in: query
                  description: 过滤表达式类型
                  schema:
                    enum:
                        - EXPR_TYPE_UNSPECIFIED
                        - AND
                        - OR
                    type: string
                    format: enum
                - name: paging.orderBy
                  in: query
                  description: 排序条件
                  schema:
                    type: string
                - name: paging.fieldMask
                  in: query
                  description: 字段掩码，其作用为SELECT中的字段，其语法为使用逗号分隔字段名，例如：id,realName,userName。如果为空则选中所有字段，即SELECT *。
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetMyTasksResponse'
    /admin/v1/oa/workflow/tasks/{taskId}:
        get:
            tags:
                - WorkflowService
            description: |-
                获取单个待办任务详情：申请标题、申请表单数据、该实例的审批日志轨迹。

                 鉴权：仅当前指派审批人可查其处于 PENDING 的任务（与 AuditTask 同款授权，
                 IDEQ + AssigneeUserIDEQ(caller) + TaskStatusEQ(PENDING)，见
                 WorkflowTaskRepo.GetDetailByAssignee）。其余字段（definition_id /
                 current_node_index / tenant_id 等）不投影，对齐 MyTaskItem 的最小披露原则。
            operationId: WorkflowService_GetTask
            parameters:
                - name: taskId
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetTaskResponse'
    /admin/v1/oa/workflow/tasks/{taskId}/audit:
        post:
            tags:
                - WorkflowService
            description: 审批/驳回/转办当前待办任务。状态机在此推进实例状态、生成下一节点任务或终结实例。
            operationId: WorkflowService_AuditTask
            parameters:
                - name: taskId
                  in: path
                  description: 绑定路径参数 {task_id}。
                  required: true
                  schema:
                    type: integer
                    format: uint32
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AuditTaskRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
...
components:
    schemas:
        AuditLogEntry:
            type: object
            properties:
                actionLabel:
                    type: string
                comment:
                    type: string
                occurredAt:
                    type: string
                    format: date-time
            description: |-
                AuditLogEntry 审批日志轨迹条目。投影自 WorkflowLog，仅含审批类动作
                 （APPROVE / REJECT / FORWARD），与 ListByActor 的过滤口径一致。
        AuditTaskRequest:
            type: object
            properties:
                taskId:
                    type: integer
                    description: 绑定路径参数 {task_id}。
                    format: uint32
                action:
                    enum:
                        - APPROVE
                        - REJECT
                        - FORWARD
                    type: string
                    format: enum
                comment:
                    type: string
                forwardToUserId:
                    type: integer
                    description: 仅 action == FORWARD 时有效：被转办人 ID。
                    format: uint32
        CreateWorkflowDefinitionRequest:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/WorkflowDefinition'
...
        GetMyTasksResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/MyTaskItem'
                total:
                    type: string
        GetTaskResponse:
            type: object
            properties:
                taskId:
                    type: integer
                    format: uint32
                instanceId:
                    type: integer
                    format: uint32
                title:
                    type: string
                formData:
                    type: string
                history:
                    type: array
                    items:
                        $ref: '#/components/schemas/AuditLogEntry'
            description: |-
                GetTaskResponse 投影待办任务详情。字段对齐详情页渲染所需，最小披露：
                 申请标题、申请表单数据（原始 JSON 文本，引擎不解释）、该实例的审批日志轨迹。
        ListWorkflowDefinitionResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/WorkflowDefinition'
                total:
                    type: string
...
        MyTaskItem:
            type: object
            properties:
                taskId:
                    type: integer
                    description: 各字段按 list_type 选择性填充。
                    format: uint32
                logId:
                    type: integer
                    format: uint32
                instanceId:
                    type: integer
                    format: uint32
                title:
                    type: string
                statusLabel:
                    type: string
                actionLabel:
                    type: string
                comment:
                    type: string
                occurredAt:
                    type: string
                    format: date-time
        SubmitApplyRequest:
            type: object
            properties:
                definitionCode:
                    type: string
                definitionVersion:
                    type: integer
                    format: uint32
                title:
                    type: string
                formData:
                    type: string
                    description: 申请表单数据，原始 JSON 文本。
        SubmitApplyResponse:
            type: object
            properties:
                instanceId:
                    type: integer
                    format: uint32
...
        WorkflowDefinition:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                name:
                    type: string
                code:
                    type: string
                version:
                    type: integer
                    format: uint32
                description:
                    type: string
                nodeConfig:
                    type: string
                    description: 节点配置，原始 JSON 文本。结构由引擎 internal/workflow 包解析。
                formSchema:
                    type: string
                    description: 动态表单 schema，原始 JSON 文本，供前端按定义渲染。
                definitionStatus:
                    enum:
                        - DRAFT
                        - ENABLED
                        - DISABLED
                    type: string
                    format: enum
                tenantId:
                    type: integer
                    description: '---- 租户 / 审计字段（与 go-wind-cms 固定字段号对齐） ----'
                    format: uint32
                tenantName:
                    type: string
                createdBy:
                    type: integer
                    format: uint32
                updatedBy:
                    type: integer
                    format: uint32
                deletedBy:
                    type: integer
                    format: uint32
                createdAt:
                    type: string
                    format: date-time
                updatedAt:
                    type: string
                    format: date-time
                deletedAt:
                    type: string
                    format: date-time
...